+++ nni/nas/evaluator/__init__.py
@@ -3,6 +3,7 @@
 
 from nni.common.framework import shortcut_framework
 
+from .evaluator import Evaluator
 from .functional import FunctionalEvaluator
 
 shortcut_framework(__name__)
